move proxy function, add 'make serve-frontend', add GITEA_DEV_FRONTEND_PORT
