+++ Makefile
@@ -188,9 +188,10 @@ help:
 	@echo " - build                            build everything"
 	@echo " - frontend                         build frontend files"
 	@echo " - backend                          build backend files"
-	@echo " - watch                            watch everything and continuously rebuild"
+	@echo " - watch                            watch backend and serve frontend"
 	@echo " - watch-frontend                   watch frontend files and continuously rebuild"
 	@echo " - watch-backend                    watch backend files and continuously rebuild"
+	@echo " - serve-frontend                   start frontend dev server"
 	@echo " - clean                            delete backend and integration files"
 	@echo " - clean-all                        delete backend, frontend and integration files"
 	@echo " - deps                             install dependencies"
@@ -423,6 +424,11 @@ watch:
 
 .PHONY: watch-frontend
 watch-frontend: node-check node_modules
+	@rm -rf $(WEBPACK_DEST_ENTRIES)
+	NODE_ENV=development npx webpack --watch --progress
+
+.PHONY: serve-frontend
+serve-frontend: node-check node_modules
 	@rm -rf $(WEBPACK_DEST_ENTRIES)
 	NODE_ENV=development npx webpack serve --progress
